refactor: add relay settings to settings and rename settings module
