+++ app/utils/files.py
@@ -4,7 +4,7 @@ import asyncio
 from datetime import UTC, datetime
 from pathlib import Path
 
-from app.core.config import settings
+from app.core.settings import settings
 
 
 async def setup_directory(path: Path) -> Path:
