+++ tests/integration/test_auth.py
@@ -4,8 +4,8 @@ import pytest
 from httpx import AsyncClient
 
 from app.auth.dependencies import create_session, reload_authorized_hashes
-from app.core.config import settings
 from app.core.runtime import AppRuntime
+from app.core.settings import settings
 
 VALID_API_KEY = "valid-key"
 AUTH_COOKIE_NAME = "relab_session"
