+++ tests/integration/test_images.py
@@ -10,7 +10,7 @@ from httpx import AsyncClient
 from pydantic import AnyUrl
 
 from app.camera.services.manager import CameraManager
-from app.core.config import settings
+from app.core.settings import settings
 from app.image_sinks.base import ImageSinkError, StoredImage
 from tests.constants import (
     QUEUED_STATUS,
